Fixing up the latest security changes and making structure better
