+++ parser_gibdd/exceptions.py
@@ -4,3 +4,11 @@ class ResourceUnreachable(Exception):
 
 class ResourceReturnedEmpty(Exception):
     pass
+
+
+class RegionNotFoundError(Exception):
+    pass
+
+
+class CrashesNotFoundError(Exception):
+    pass
