+++ parser_gibdd/models/region.py
@@ -1,10 +1,20 @@
 import logging
-from typing import List, Optional, Union
+from typing import Optional, List, Union
 
 from fuzzywuzzy import fuzz
 from pydantic import BaseModel
 
+from exceptions import RegionNotFoundError
+
 logger = logging.getLogger(__name__)
+
+
+class RegionName(str):
+    pass
+
+
+class FederalRegionName(str):
+    pass
 
 
 class Region(BaseModel):
@@ -50,4 +60,4 @@ class Country(Region):
                 return federal
             if region.okato in [inner.okato for inner in federal.districts]:
                 return federal
-        raise Exception("region not found")
+        raise RegionNotFoundError(f"Region {region.name} not found in {self.name}")
