+++ pyproject.toml
@@ -11,14 +11,14 @@ packages = [{ include = "src" }]
 [tool.poetry.dependencies]
 python = "^3.8"
 click = "^7.1.2"
-pydantic = "^1.6.1"
+pydantic = "^1.6.2"
 requests = "^2.24.0"
 python-dotenv = "^0.14.0"
 pandas = "^1.1.2"
 xlrd = "^1.2.0"
 openpyxl = "^3.0.5"
 fuzzywuzzy = "^0.18.0"
-python-Levenshtein = "^0.12.0"
+Levenshtein = "^0.16.0"
 
 [tool.poetry.dev-dependencies]
 pytest = "^6.1.1"
@@ -26,7 +26,7 @@ setuptools = "^50.3.0"
 mypy = "^0.782"
 
 [tool.poetry.scripts]
-gibdd = "src.gibdd_cli:main"
+gibdd = "src.cli.shell:main"
 
 [build-system]
 requires = ["poetry_core>=1.0.0"]
