+++ requirements.txt
@@ -1,18 +1,20 @@
-certifi==2020.6.20
-chardet==3.0.4
-click==7.1.2
-et-xmlfile==1.0.1
+certifi==2021.10.8; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
+charset-normalizer==2.0.9; python_full_version >= "3.6.0" and python_version >= "3"
+click==7.1.2; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
+et-xmlfile==1.1.0; python_version >= "3.6"
 fuzzywuzzy==0.18.0
-idna==2.10
-jdcal==1.4.1
-numpy==1.19.2
-openpyxl==3.0.5
-pandas==1.1.2
-pydantic==1.6.1
-python-dateutil==2.8.1
+idna==3.3; python_version >= "3.5" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.5"
+levenshtein==0.16.0; python_version >= "3.5"
+numpy==1.21.1
+openpyxl==3.0.9; python_version >= "3.6"
+pandas==1.3.4; python_full_version >= "3.7.1"
+pydantic==1.8.2; python_full_version >= "3.6.1"
+python-dateutil==2.8.2; python_full_version >= "3.7.1"
 python-dotenv==0.14.0
-pytz==2020.1
-requests==2.24.0
-six==1.15.0
-urllib3==1.25.10
-xlrd==1.2.0
+pytz==2021.3; python_full_version >= "3.7.1"
+rapidfuzz==1.8.3; python_version >= "3.5"
+requests==2.26.0; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.6.0")
+six==1.16.0; python_full_version >= "3.7.1"
+typing-extensions==4.0.1; python_version >= "3.6" and python_full_version >= "3.6.1"
+urllib3==1.26.7; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version < "4"
+xlrd==1.2.0; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.4.0")
