--- src/models/region.py
@@ -1,5 +1,0 @@
-from typing import TypeVar
-
-
-RegionName = TypeVar("RegionName", str, str)
-FederalRegionName = TypeVar("FederalRegionName", str, str)
